Create main.py
Dictionary Builder is a custom project of mine with a graphical user interface that can be used to create a custom set of words and meanings. It can save and load files in json format.
I have also made sure that it doesn't use any invalid file. This app may help those who want to create a fantasy based language with unique words and meanings.
This app uses tkinter, os, re and json module. 

Features 
-Can store multiple meanings (synonyms) for one word
-Translates a word and shows it primary meaning
-Only the first set meaning is considered to be the primary meaning
-Can check the number of words
-The primary meaning method for now only works with default "Language 1"
-It is not required to input language names, however file name must be defined
-Detailed instructions will be given latter in a future update

Features planned to install
-Delete or modify word meanings
-Translate strings or documents 
-Better alternative to primary meaning method